+++ src/main/java/com/won/myongjiCamp/config/security/CustomAuthenticationEntryPoint.java
@@ -1,0 +1,26 @@
+package com.won.myongjiCamp.config.security;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.won.myongjiCamp.dto.response.ResponseDto;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response,
+                         AuthenticationException authException) throws IOException, ServletException {
+
+        response.setContentType("application/json;charset=UTF-8");
+        response.setStatus(HttpStatus.UNAUTHORIZED.value());
+
+        response.getWriter().write(new ObjectMapper().writeValueAsString(
+                new ResponseDto<>(response.getStatus(), "인증이 필요한 요청입니다.")));
+    }
+}
