+++ src/main/java/com/won/myongjiCamp/config/security/auth/PrincipalDetailService.java
@@ -1,4 +1,4 @@
-package com.won.myongjiCamp.config.auth;
+package com.won.myongjiCamp.config.security.auth;
 
 import com.won.myongjiCamp.model.Member;
 import com.won.myongjiCamp.repository.MemberRepository;
