+++ src/main/java/com/won/myongjiCamp/controller/api/ResumeApiController.java
@@ -1,6 +1,6 @@
 package com.won.myongjiCamp.controller.api;
 
-import com.won.myongjiCamp.config.auth.PrincipalDetail;
+import com.won.myongjiCamp.config.security.auth.PrincipalDetail;
 import com.won.myongjiCamp.dto.response.ResponseDto;
 import com.won.myongjiCamp.dto.request.ResumeDto;
 import com.won.myongjiCamp.exception.MemberNoMatchException;
